pyproject.toml
[project]
name = "ankiany"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "claude-agent-sdk>=0.1.17",
    "genanki>=0.13.1",
]
